type declarations

+++ src/main/flight/pid.h
@@ -160,7 +160,7 @@ typedef struct pidAutotuneConfig_s {
     uint8_t     fw_max_rate_threshold;      // Threshold [%] of max rate to consider autotune detection
     uint8_t     fw_ff_to_p_gain;            // FF to P gain (strength relationship) [%]
     uint16_t    fw_ff_to_i_time_constant;   // FF to I time (defines time for I to reach the same level of response as FF) [ms]
-    uint16_t    fw_autotune_rate_adjustment;  // Target rate adjustment on full stick deflection?
+    uint8_t     fw_autotune_rate_adjustment;  // Target rate adjustment on full stick deflection?
 } pidAutotuneConfig_t;
 
 typedef enum {
